hibernate dependency fixed

+++ src/main/java/com/example/g_ore/MyUserDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MyUserDAO extends JpaRepository<MyUser, Long> {
+public interface MyUserDAO extends JpaRepository<MyUser, Integer> {
 
 }
